Link to project local .npmrc
Running the `npm install` command in the `INIT_CWD` doesn't work since the location that files get installed to is not the `INIT_CWD` directory until install finishes. That means that the .dat file that is supposed to get copied must be a relative path.

To fix this, we can link to the .npmrc in `INIT_CWD` run the install and copy the .dat file relatively.